pihole/requirements.txt
gmqtt
pihole6api
colorlog
python-logging-loki